Add new GitHub Actions workflow for updating "dist" branches
This workflow is manually triggered by going to the "Actions" tab, clicking on the appropriate workflow in the sidebar, and choosing the option to "Run" it (against the default branch).

This allows it to not only be managed more effectively (directly in GitHub) but also makes it run significantly faster since all architectures can update simultaneously.